server folder structure changed
