--- packages/server/src/auth/jwt-payload.interface.ts
@@ -1,3 +1,0 @@
-export interface JwtPayload {
-  email: string;
-}
