--- packages/server/src/auth/jwt.strategy.ts
@@ -1,32 +1,0 @@
-import { Injectable, UnauthorizedException } from '@nestjs/common';
-import { PassportStrategy } from '@nestjs/passport';
-import { ExtractJwt, Strategy } from 'passport-jwt';
-import { UserRepository } from './user.repository';
-import { JwtPayload } from './jwt-payload.interface';
-import { User } from './user.entity';
-import { ConfigService } from '@nestjs/config';
-
-@Injectable()
-export class JwtStrategy extends PassportStrategy(Strategy) {
-  constructor(
-    private userRepository: UserRepository,
-    private configService: ConfigService,
-  ) {
-    super({
-      secretOrKey: configService.get('JWT_SECRET'),
-      jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
-    });
-  }
-
-  async validate(payload: JwtPayload) {
-    const { email } = payload;
-
-    const user: User = await this.userRepository.findOneBy({ email });
-
-    if (!user) {
-      throw new UnauthorizedException();
-    }
-
-    return user;
-  }
-}
