+++ packages/server/src/modules/auth/jwt-payload.interface.ts
@@ -1,0 +1,3 @@
+export interface JwtPayload {
+  email: string;
+}
