+++ packages/server/src/modules/auth/user.entity.ts
