+++ packages/server/src/modules/listing/listing.module.ts
@@ -5,9 +5,9 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { Listing } from './entities/listing.entity';
 import { AuthModule } from '../auth/auth.module';
 import { ListingRepository } from './listing.repository';
-import { UploadModule } from '../common/upload/upload.module';
 import { ListingImage } from './entities/listing-image.entity';
 import { ListingImageService } from './services/listing-image.service';
+import { UploadModule } from '../../shared/upload/upload.module';
 
 @Module({
   imports: [
